+++ src/main/java/com/wisr/mlsched/job/SuperLinearLossFunction.java
@@ -3,12 +3,12 @@ package com.wisr.mlsched.job;
 import java.util.Random;
 
 public class SuperLinearLossFunction implements LossFunction {
-  public int numIterations;
+  public long numIterations;
   public double valueBegin;
   public double valueEnd;
   public double u, b, c;
 
-  private SuperLinearLossFunction(int numIterations, double valueBegin, double valueEnd,
+  private SuperLinearLossFunction(long numIterations, double valueBegin, double valueEnd,
                            double u, double b, double c) {
     this.numIterations = numIterations;
     this.valueBegin = valueBegin;
@@ -18,7 +18,7 @@ public class SuperLinearLossFunction implements LossFunction {
     this.c = c;
   }
 
-  public static SuperLinearLossFunction getRandomSuperlinearFunction(int numIterations, int seed) {
+  public static SuperLinearLossFunction getRandomSuperlinearFunction(long numIterations, int seed) {
     double tValueBegin = 1.0;
     double tValueEnd = 0.0;
     Random r = new Random(seed);
@@ -30,18 +30,18 @@ public class SuperLinearLossFunction implements LossFunction {
         tU, tB, tC);
   }
 
-  public double getValue(int iteration) {
+  public double getValue(long iteration) {
     double value = Math.pow(u, iteration - b) + c;
     return value;
   }
 
-  public double getSlope(int iteration) {
+  public double getSlope(long iteration) {
     double value = Math.pow(u, iteration - b);
     value = Math.log(u)*value;
     return value;
   }
 
-  public double getDeltaValue(int iteration) {
+  public double getDeltaValue(long iteration) {
     return getValue(iteration-1) - getValue(iteration);
   }
 }
